CMake: MPI, Asserts and Advanced AMReX (#44)
* CMake: HiPACE_MPI

Wrap the last user-facing AMReX option into a HiPACE option. This
reduces confusion and improves consistency further.

* CMake: Assertations

Enable AMReX' `ENABLE_ASSERTIONS` with build type `Debug` and allow
to overwrite on the command line with `-DENABLE_ASSERTIONS=ON` for
manual activation with other build types.

* CMake: Floating Point Exceptions

Enable AMReX' `ENABLE_FPE` with build type `Debug` and allow
to overwrite on the command line with `-DENABLE_FPE=ON` for
manual activation with other build types.

* CMake: Toggle for CMAKE_BUILD_TYPE

Add allowed options for toggling the `CMAKE_BUILD_TYPE`.

* CMake: Advanced AMReX Options

Declutter user-facing options further.

* CMake: AMReX Advanced Options

For some advanced AMReX options, we want to avoid setting them with
`INTERNAL` as this implies `FORCE` and we cannot overwrite them from
the CMake GUIs.