CMake: DPC++, OMP Def., Binary Names (#47)
With regards to the `HiPACE_COMPUTE` options the following changes
are introduced:
- default: `OMP` (OpenMP; as in GNUmake and a sensible default for
                  reasonable performance in 2020)
- rename: `NONE` -> `NOACC` (no accelerated computing backend)
- adds: `DPCPP` (for SYCL/Intel DPC++ beta7 or newer)

Furthermore, we encode relevant, user-facing options in our binary
name. The namings are intentionally a bit different now and skip few
options that are verbose and standard. Furthermore, this builds a
`hipace` symlink in `bin/` to the latest build executable (when
switching options), so that repeated compile-and-test workflows do
not need to take care of the executable name anymore.