CMake: New AMReX Options (#204)
* CMake: New AMReX Options

Anticipate new AMReX options in CMake.

Also introduces a new environment hint, `AMReX_CUDA_ARCH`, that we
can use in `hipace.profile` files to default-optimize for CUDA
architectures. Currently, as well as without the hint, CMake will
try to find out local GPUs by running a small executable that queries
the device capabilities. If the login node has no GPUs or the run
does not succeed we fall back to potentially unoptimized default
architectures (6.0+).

* CMake 3.18+: CUDA Arch Policy (OLD)

Keep the old policy for now to avoid setting the code generation
flags twice and silence a warning.

https://cmake.org/cmake/help/latest/policy/CMP0104.html

We should be able to transition this well, but I need to first find
out how device LTO generation flags are handled here, if at all,
and how to restore the default-detection of the local GPU
architecture, e.g. on Summits head-nodes.

* CI: CUDA Test

* Comment Tests